fix: overhaul how we spawn commands under seccomp/landlock on Linux
