+++ codex-rs/cli/src/main.rs
@@ -1,15 +1,10 @@
-use std::path::PathBuf;
-
 use clap::Parser;
 use codex_cli::LandlockCommand;
 use codex_cli::SeatbeltCommand;
-use codex_cli::create_sandbox_policy;
 use codex_cli::proto;
-use codex_cli::seatbelt;
-use codex_core::config::Config;
-use codex_core::config::ConfigOverrides;
 use codex_exec::Cli as ExecCli;
 use codex_tui::Cli as TuiCli;
+use std::path::PathBuf;
 
 use crate::proto::ProtoCli;
 
@@ -66,14 +61,14 @@ enum DebugCommand {
 #[derive(Debug, Parser)]
 struct ReplProto {}
 
-#[tokio::main]
-async fn main() -> anyhow::Result<()> {
-    let codex_linux_sandbox_exe: Option<PathBuf> = if cfg!(target_os = "linux") {
-        std::env::current_exe().ok()
-    } else {
-        None
-    };
+fn main() -> anyhow::Result<()> {
+    codex_linux_sandbox::run_with_sandbox(|codex_linux_sandbox_exe| async move {
+        cli_main(codex_linux_sandbox_exe).await?;
+        Ok(())
+    })
+}
 
+async fn cli_main(codex_linux_sandbox_exe: Option<PathBuf>) -> anyhow::Result<()> {
     let cli = MultitoolCli::parse();
 
     match cli.subcommand {
@@ -90,34 +85,11 @@ async fn main() -> anyhow::Result<()> {
             proto::run_main(proto_cli).await?;
         }
         Some(Subcommand::Debug(debug_args)) => match debug_args.cmd {
-            DebugCommand::Seatbelt(SeatbeltCommand {
-                command,
-                sandbox,
-                full_auto,
-            }) => {
-                let sandbox_policy = create_sandbox_policy(full_auto, sandbox);
-                let config = Config::load_with_overrides(ConfigOverrides {
-                    sandbox_policy: Some(sandbox_policy),
-                    ..Default::default()
-                })?;
-                seatbelt::run_seatbelt(command, &config).await?;
+            DebugCommand::Seatbelt(seatbelt_command) => {
+                codex_cli::debug_sandbox::run_command_under_seatbelt(seatbelt_command).await?;
             }
-            #[cfg(unix)]
-            DebugCommand::Landlock(LandlockCommand {
-                command,
-                sandbox,
-                full_auto,
-            }) => {
-                let sandbox_policy = create_sandbox_policy(full_auto, sandbox);
-                let config = Config::load_with_overrides(ConfigOverrides {
-                    sandbox_policy: Some(sandbox_policy),
-                    ..Default::default()
-                })?;
-                codex_cli::landlock::run_landlock(command, &config)?;
-            }
-            #[cfg(not(unix))]
-            DebugCommand::Landlock(_) => {
-                anyhow::bail!("Landlock is only supported on Linux.");
+            DebugCommand::Landlock(landlock_command) => {
+                codex_cli::debug_sandbox::run_command_under_landlock(landlock_command).await?;
             }
         },
     }
