+++ codex-rs/core/Cargo.toml
@@ -49,7 +49,6 @@ uuid = { version = "1", features = ["serde", "v4"] }
 wildmatch = "2.4.0"
 
 [target.'cfg(target_os = "linux")'.dependencies]
-libc = "0.2.172"
 landlock = "0.4.1"
 seccompiler = "0.5.0"
