+++ codex-rs/exec/src/main.rs
@@ -1,19 +1,22 @@
-use std::path::PathBuf;
-
+//! Entry-point for the `codex-exec` binary.
+//!
+//! When this CLI is invoked normally, it parses the standard `codex-exec` CLI
+//! options and launches the non-interactive Codex agent. However, if it is
+//! invoked with arg0 as `codex-linux-sandbox`, we instead treat the invocation
+//! as a request to run the logic for the standalone `codex-linux-sandbox`
+//! executable (i.e., parse any -s args and then run a *sandboxed* command under
+//! Landlock + seccomp.
+//!
+//! This allows us to ship a completely separate set of functionality as part
+//! of the `codex-exec` binary.
 use clap::Parser;
 use codex_exec::Cli;
 use codex_exec::run_main;
 
-#[tokio::main]
-async fn main() -> anyhow::Result<()> {
-    let codex_linux_sandbox_exe: Option<PathBuf> = if cfg!(target_os = "linux") {
-        std::env::current_exe().ok()
-    } else {
-        None
-    };
-
-    let cli = Cli::parse();
-    run_main(cli, codex_linux_sandbox_exe).await?;
-
-    Ok(())
+fn main() -> anyhow::Result<()> {
+    codex_linux_sandbox::run_with_sandbox(|codex_linux_sandbox_exe| async move {
+        let cli = Cli::parse();
+        run_main(cli, codex_linux_sandbox_exe).await?;
+        Ok(())
+    })
 }
