+++ codex-rs/mcp-server/src/main.rs
@@ -1,15 +1,8 @@
-use std::path::PathBuf;
-
 use codex_mcp_server::run_main;
 
-#[tokio::main]
-async fn main() -> std::io::Result<()> {
-    let codex_linux_sandbox_exe: Option<PathBuf> = if cfg!(target_os = "linux") {
-        std::env::current_exe().ok()
-    } else {
-        None
-    };
-
-    run_main(codex_linux_sandbox_exe).await?;
-    Ok(())
+fn main() -> anyhow::Result<()> {
+    codex_linux_sandbox::run_with_sandbox(|codex_linux_sandbox_exe| async move {
+        run_main(codex_linux_sandbox_exe).await?;
+        Ok(())
+    })
 }
